Removed priority field for tasks
Removed priority field for each task, because
it can confuse the user, he already has stages
and there is no need to have another entity,
to define the order of several tasks within
single stage we can the order which the tasks
were defined in, e.g. Ansible uses the same
way, to define the order of tasks.

Change-Id: Ie0539266a0339b2bbf2857c88694d0a691338ec6
Implements: blueprint cinder-neutron-plugins-in-fuel