Explicitly specify 'puppet_modules' for tasks
In Astute, we have a strict validation that 'puppet_modules' must
present in puppet task declaration; otherwise fail deployment. That
means 'puppet_modules' is not allowed to be an empty string.

This validation is presented only for pre/post deployment tasks, and
isn't performed for deployment tasks. But let's be pessimistic, and
force users to specify 'puppet_modules' explicitly even if it's
current working directory.

Change-Id: I63e546545077bea15f62c8ac8682c682895e5935
Closes-Bug: #1506624