Allow arrays in 'value' in environment_config.yaml
Since Fuel 9.0 we have a new type for environment attributes: "text_list".
In that case, attribute's value must be an array. FPB didn't handle that
case, and throws error for any case when value is array.

This patch add "array" as a valid type for "value". In order to keep
simplicity and do not overcomplicate JSON schema, it allows arrays for
all types, not only for "text_list". It's consistent with Nailgun code,
where we don't have such check either.

Change-Id: I2ab401e4cdc7c1a1eb276aa71253818c688fdca1
Closes-Bug: #1616466