Legacy tasks are turned back for exmaple v3.0.0
In previous patch tasks.yaml contained legacy tasks was cleaned up
but it turned up that some QA tests was relying on this tasks output
so this tasks was moved to the deployment_tasks.yaml to provide this
output.

Double unfortuantely FPB validation does not support tasks <v1.0.0
validation for the deployment_tasks.yaml without major changes in
validation logic.

So the tasks was converted to the tasks v1.0.0 by adding tasks id.

Change-Id: I1f820030a0575155b06db3c493accadc252c241a
Closes-Bug: #1625244